% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_index.R
\name{add_index}
\alias{add_index}
\title{Add an index to a SQL table}
\usage{
add_index(
  conn,
  server = NULL,
  config = NULL,
  config_url = NULL,
  config_file = NULL,
  to_schema = NULL,
  to_table = NULL,
  index_type = NULL,
  index_name = NULL,
  index_vars = NULL,
  index_ts = T,
  drop_index = T,
  test_schema = NULL
)
}
\arguments{
\item{conn}{SQL server connection created using \code{odbc} package.}

\item{server}{Name of server being used (only applies if using a YAML file).
Useful if the same table is loaded to multiple servers but with different names
or schema.}

\item{config}{Name of object in global environment that contains configuration
information. Use one of \code{config}, \code{config_url}, or \code{config_file}.
\if{html}{\out{<br>}}\if{html}{\out{<br>}}
Should be in a YAML format with at least the following variables: \strong{to_schema}, \strong{to_table}, and \strong{index_name}.
\if{html}{\out{<br>}}\if{html}{\out{<br>}}
The \code{config} \emph{may} also contain \strong{index_type} and \strong{index_vars} variables.
\if{html}{\out{<br>}}\if{html}{\out{<br>}}
\strong{to_schema} and \strong{to_table} should be nested under the server name if applicable,
other variables should not.
\if{html}{\out{<br>}}\if{html}{\out{<br>}}
If \strong{index_type} is not specified and \strong{index_vars}
fields are named, the assumption is a clustered (CL) index.}

\item{config_url}{URL of a YAML config file. Use one of \code{config}, \code{config_url}, or
\code{config_file}. Note the requirements under \code{config}.}

\item{config_file}{File path of a YAML config file. Use one of \code{config}, \code{config_url}, or
\code{config_file}. Note the requirements under \code{config}.}

\item{to_schema}{Name of the schema to apply the index to (if not using YAML input).}

\item{to_table}{Name of the table to apply the index to (if not using YAML input).}

\item{index_type}{Which index type will be used, either 'ccs' or 'cl' (if not using YAML input).}

\item{index_name}{Name of the index to be added (if not using YAML input).}

\item{index_vars}{Vector of variables to index on if using a clustered ('cl') index (if not using YAML input).
Should take the format \code{c("a", "b", "c")}.}

\item{index_ts}{This will append a timestamp in the format of YYYYMMDD-HHMMSS to
the \code{index_name}, guaranteeing uniqueness. Default is \code{TRUE}.}

\item{drop_index}{Remove any existing clustered or clustered column store indices.
Default is \code{TRUE}.}

\item{test_schema}{Add index to a temporary/development schema when testing out table creation.
Will use the to_schema (specified or in the YAML file) to make a new table name of
\verb{\{to_schema\}_\{to_table\}}. Schema must already exist in the database. Most useful
when the user has an existing YAML file and does not want to overwrite it. Default is \code{NULL}.}
}
\value{
No return value. Function is called for its side effects of adding an index to a SQL table.
}
\description{
\code{add_index()} adds an index to a SQL table using specified variables or a YAML config file.
}
\details{
This function adds a clustered column store (CCS) or clustered (CL)
index to a SQL table using specified variables or a YAML configuration file.
Users can specify some input functions (e.g., to_table) and rely on the config
file for the rest of the necessary information. For all arguments that could be
specified or come from a YAML file, the hierarchy is:
specified > argument under server in YAML > argument not under server in YAML.
\subsection{Example YAML file with no server and a CCS index}{

(Assume the indentation is appropriate)

\if{html}{\out{<div class="sourceCode yaml">}}\preformatted{to_schema: raw
to_table: mcaid_elig
# optional other components like a qa_schema and qa_table, variables, etc.
index_type: ccs
index_name: idx_ccs_raw_mcaid_elig
}\if{html}{\out{</div>}}
}

\subsection{Example YAML file with servers (phclaims, hhsaw) and clustered index}{

(Assume the indentation is appropriate)

\if{html}{\out{<div class="sourceCode yaml">}}\preformatted{phclaims:
    to_schema: raw
    to_table: mcaid_elig
hhsaw:
    to_schema: raw
    to_table: mcaid_elig
# optional other components like a qa_schema and qa_table, variables, etc.
index_type: cl
index_name: idx_raw_mcaid_elig_id_date
index_vars:
    - id_apde
    - start_date
}\if{html}{\out{</div>}}
}
}
\note{
This function replaces the deprecated \code{add_index} function from the
\code{apde} package.
}
\examples{
\dontrun{
add_index(conn = db_claims, server = "hhsaw", config = load_config)
}

}
