% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apde_keyring.R
\name{apde_keyring_check}
\alias{apde_keyring_check}
\title{Check and Create Keyring if Needed}
\usage{
apde_keyring_check(keyring)
}
\arguments{
\item{keyring}{A character string specifying the name of the keyring.}
}
\value{
A character message indicating whether the keyring exists or has been created.
}
\description{
This function checks if a keyring exists and creates it if it does not.
}
\note{
This function replaces the deprecated \code{apde_keyring_check_f} function from the
\code{apde} package.
}
\examples{
\dontrun{
apde_keyring_check("my_keyring")
}

}
