% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apde_keyring.R
\name{apde_keyring_set}
\alias{apde_keyring_set}
\title{Set or Update a Keyring}
\usage{
apde_keyring_set(keyring = NA)
}
\arguments{
\item{keyring}{A character string specifying the name of the keyring. If not provided, a dialog will prompt for input.}
}
\value{
A character message indicating the keyring and username that have been set.
}
\description{
This function creates or updates a keyring with a specified username.
}
\note{
This function replaces the deprecated \code{apde_keyring_set_f} function from the
\code{apde} package.
}
\examples{
\dontrun{
apde_keyring_set("my_keyring")
}

}
