% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.R
\name{apde_notify_address_create}
\alias{apde_notify_address_create}
\title{Create a New APDE Notify Address}
\usage{
apde_notify_address_create(address)
}
\arguments{
\item{address}{Character. Email address to create.}
}
\value{
Integer. ID of the newly created address.
}
\description{
\strong{Internal function}. Creates a new email address in the system.
}
\details{
The function checks that the address doesn't already exist before creating it.
}
\examples{
 \dontrun{
  new_id <- apde_notify_address_create(address = "user@kingcounty.gov")
 }
}
\keyword{internal}
