% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.R
\name{apde_notify_address_delete}
\alias{apde_notify_address_delete}
\title{Delete an APDE Notify Address}
\usage{
apde_notify_address_delete(address_id = NULL, address = NULL)
}
\arguments{
\item{address_id}{Integer. ID of the address to delete.}

\item{address}{Character. Email address to delete.}
}
\value{
None
}
\description{
\strong{Internal function}. Deletes an email address and all associated list memberships.
}
\details{
Either address_id or address must be provided. This function will also remove
the address from all notification lists before deleting the address record.
}
\examples{
 \dontrun{
  apde_notify_address_delete(address = "user@kingcounty.gov")
 }
}
\keyword{internal}
