% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.R
\name{apde_notify_address_get}
\alias{apde_notify_address_get}
\title{Retrieve a Specific APDE Notify Address}
\usage{
apde_notify_address_get(address_id = NULL, address = NULL)
}
\arguments{
\item{address_id}{Integer. ID of the address to retrieve.}

\item{address}{Character. Email address to retrieve.}
}
\value{
A data.frame containing the address information.
}
\description{
\strong{Internal function}. Retrieves a specific email address by ID or address value.
}
\details{
Either address_id or address must be provided. The function will use
apde_notify_address_id_get() to resolve the ID if only address is provided.
}
\examples{
 \dontrun{
  address_info <- apde_notify_address_get(address = "user@kingcounty.gov")
 }
}
\keyword{internal}
