% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.R
\name{apde_notify_address_set}
\alias{apde_notify_address_set}
\title{Update an APDE Notify Address}
\usage{
apde_notify_address_set(address_id = NULL, address = NULL, new_address)
}
\arguments{
\item{address_id}{Integer. ID of the address to update.}

\item{address}{Character. Current email address.}

\item{new_address}{Character. New email address.}
}
\value{
None
}
\description{
\strong{Internal function}. Updates an existing email address in the system.
}
\details{
Either address_id or address must be provided to identify the record to update.
The function checks that the new_address doesn't already exist before updating.
}
\examples{
 \dontrun{
  apde_notify_address_set(address = "old@kingcounty.gov",
                            new_address = "new@kingcounty.gov")
 }
}
\keyword{internal}
