% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.R
\name{apde_notify_addresses_get}
\alias{apde_notify_addresses_get}
\title{Retrieve All APDE Notify Addresses}
\usage{
apde_notify_addresses_get()
}
\value{
A data.frame containing all notify addresses with columns: id, address
}
\description{
\strong{Internal function}. Retrieves all email addresses from the notify_addresses table.
}
\details{
This function connects to the HHSAW database and returns all email addresses
stored in the apde.notify_addresses table, ordered alphabetically by address.
}
\examples{
 \dontrun{
  addresses <- apde_notify_addresses_get()
 }
}
\keyword{internal}
