% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.R
\name{apde_notify_list_set}
\alias{apde_notify_list_set}
\title{Set APDE Notify List for a Message}
\usage{
apde_notify_list_set(msg_id = NULL, msg_name = NULL, choices)
}
\arguments{
\item{msg_id}{Integer. ID of the message.}

\item{msg_name}{Character. Name of the message.}

\item{choices}{Vector. List of email addresses to associate with the message.}
}
\value{
None
}
\description{
\strong{Internal function}. Replaces the email address list for a specific message.
}
\details{
Either msg_id or msg_name must be provided. This function completely replaces
the existing recipient list with the provided choices.
}
\examples{
 \dontrun{
  apde_notify_list_set(msg_name = "daily_report",
                         choices = c("user1@kingcounty.gov", "user2@kingcounty.gov"))
 }
}
\keyword{internal}
