% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.R
\name{apde_notify_msg_set}
\alias{apde_notify_msg_set}
\title{Set or Update APDE Notify Message}
\usage{
apde_notify_msg_set(msg_id = 0, msg_name, msg_subject, msg_body, msg_from)
}
\arguments{
\item{msg_id}{Integer. ID of the message to update. Use 0 for new messages.}

\item{msg_name}{Character. Name of the message.}

\item{msg_subject}{Character. Subject of the message.}

\item{msg_body}{Character. Body of the message.}

\item{msg_from}{Character. Sender's email address.}
}
\value{
Integer. ID of the new or updated message.
}
\description{
\strong{Internal function}. Creates a new message template or updates an existing one.
}
\details{
If msg_id is 0, creates a new message. If msg_id > 0, creates a new version
and archives the old one by setting its msg_parent field. This maintains
a history of message changes.
}
\examples{
 \dontrun{
  new_id <- apde_notify_msg_set(msg_id = 0,
                                  msg_name = "test_message",
                                  msg_subject = "Test Subject",
                                  msg_body = "Test body",
                                  msg_from = "sender@kingcounty.gov")
 }
}
\seealso{
\code{\link[=apde_notify_menu]{apde_notify_menu()}} which calls this function

\code{\link[=apde_notify_msg_get]{apde_notify_msg_get()}} which is called by this function
}
\keyword{internal}
