% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_into.R
\name{copy_into}
\alias{copy_into}
\title{Copy Data from the Data Lake to the Data Warehouse}
\usage{
copy_into(
  conn,
  server = NULL,
  config = NULL,
  config_url = NULL,
  config_file = NULL,
  to_schema = NULL,
  to_table = NULL,
  db_name = NULL,
  dl_path = NULL,
  file_type = c("csv", "parquet", "orc"),
  identity = NULL,
  secret = NULL,
  max_errors = 100,
  compression = c("none", "gzip", "defaultcodec", "snappy"),
  field_quote = "",
  field_term = NULL,
  row_term = NULL,
  first_row = 2,
  overwrite = TRUE,
  rodbc = FALSE,
  rodbc_dsn = "int_edw_16"
)
}
\arguments{
\item{conn}{SQL server connection created using \code{odbc} package}

\item{server}{Server name: 'hhsaw' or 'phclaims'. If NULL, will be set to NA.}

\item{config}{A list object in memory containing the YAML config file contents (should be NULL if using config_url or config_file)}

\item{config_url}{The URL location of the YAML config file (should be NULL if using config or config_file)}

\item{config_file}{The file path of the YAML config file (should be NULL if using config or config_url)}

\item{to_schema}{Target schema name in the data warehouse}

\item{to_table}{Target table name in the data warehouse}

\item{db_name}{Database name (e.g., "hhs_analytics_workspace", "inthealth_edw")}

\item{dl_path}{The data lake path where the source files are located}

\item{file_type}{File type: \code{"csv"}, \code{"parquet"}, or \code{"orc"}}

\item{identity}{The identity (username or account name) for data lake authentication}

\item{secret}{The secret key or password for data lake authentication}

\item{max_errors}{Maximum number of records that can be rejected before the entire file is rejected (default: \code{100})}

\item{compression}{Compression type: \code{"none"}, \code{"gzip"}, \code{"defaultcodec"}, or \code{"snappy"} (default: \code{"none"})}

\item{field_quote}{Character used to quote fields in the input file (default: empty string)}

\item{field_term}{Character or string used to separate fields in the input file}

\item{row_term}{Character or string used to separate rows in the input file}

\item{first_row}{Row number where data begins in the input file, excluding headers (default: \code{2})}

\item{overwrite}{Logical; if TRUE, truncate the table before creating it (default: \code{TRUE})}

\item{rodbc}{Logical; if TRUE, use RODBC package to avoid encoding errors with secret keys (default: \code{FALSE})}

\item{rodbc_dsn}{DSN name for the RODBC connection (default: \code{"int_edw_16"})}
}
\value{
None (invisible NULL)
}
\description{
This function copies data from the data lake to the data warehouse using SQL Server's COPY INTO statement.
}
\details{
Refactored on 7/25/2025 by Danny colombara to improve modularity and maintainability
by extracting helper functions and improving code organization.

Plans for future improvements:
\itemize{
\item Add warning when table is about to be overwritten.
\item Add other options for things we're not using (e.g., file_format).
}
}
\note{
This function replaces the deprecated \code{copy_into_f} function from the
\code{apde} package. The functionality is identical, but this version has
improved code organization and maintainability.
}
\author{
Alastair Matheson, 2019-04-04
}
