% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_into.R
\name{create_auth_sql}
\alias{create_auth_sql}
\title{HELPER: Create authentication SQL fragment}
\usage{
create_auth_sql(rodbc, identity, secret, conn)
}
\arguments{
\item{rodbc}{Whether using RODBC}

\item{identity}{Authentication identity}

\item{secret}{Authentication secret}

\item{conn}{Database connection}
}
\value{
SQL fragment for authentication
}
\description{
HELPER: Create authentication SQL fragment
}
