% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_db_connection.R
\name{create_db_connection}
\alias{create_db_connection}
\title{Create Database Connection Function}
\usage{
create_db_connection(
  server = c("phextractstore", "hhsaw", "inthealth"),
  prod = TRUE,
  interactive = FALSE
)
}
\arguments{
\item{server}{Whether we are connecting to \code{'hhsaw'}, \code{'inthealth'},
or \code{'phextractstore'}}

\item{prod}{Whether the production or dev/WIP servers are to be used. Defaults
to \code{prod = TRUE}.}

\item{interactive}{Whether to input password manually for HHSAW/inthealth or
to use keyring. \code{TRUE} should be used when connecting on VM, typically,
or if keyring/odbc are not set up yet for a given server. Defaults to
\code{interactive = FALSE}.}
}
\value{
A database connection object
}
\description{
Create a connection to APDE prod or dev servers
}
\note{
This function replaces the deprecated \code{create_db_connection} function from the
\code{apde} package.
}
\examples{
\dontrun{
# Connect to production HHSAW
conn <- create_db_connection(server = "hhsaw", prod = TRUE)

# Connect to development inthealth server
conn <- create_db_connection(server = "inthealth", prod = FALSE)

# Connect to phextractstore with interactive authentication
conn <- create_db_connection(server = "phextractstore", interactive = TRUE)
}

}
\author{
Jeremy Whitehurst

Kai Fukutaki
}
