% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_qa_run_pipeline.R
\name{etl_qa_export_results}
\alias{etl_qa_export_results}
\title{Export tables and graphs of ETL QA pipeline results}
\usage{
etl_qa_export_results(qa_results, config)
}
\arguments{
\item{qa_results}{A list containing the processed QA results from \code{etl_qa_final_results()}}

\item{config}{An S3 object of class "qa_data_config" containing configuration settings.}
}
\value{
A list the file paths for the exported data:
\itemize{
\item \code{pdf_missing}: File path to PDF of plots of data missingness over time
\item \code{pdf_values}: File path to PDF of plots of frequency and statistical changes over time
\item \code{excel}: File path to Excel file with tabs for missingness, values, CHI comparisons
}
}
\description{
This function exports Excel tables and PDF plots of ETL QA results. It is the
fourth and final step run by \code{\link[=etl_qa_run_pipeline]{etl_qa_run_pipeline()}}.
}
\details{
This is an \emph{internal function} accessible only by use of \code{:::}, for example,
\code{apde.etl:::etl_qa_export_results(...)}.
}
\examples{
\dontrun{

# Step 1: generate a config object
myconfig <- etl_qa_setup_config(
  data_source_type = 'rads',
  data_params = list(
    function_name = 'get_data_birth',
    time_var = 'chi_year',
    time_range = c(2021, 2022),
    cols = c('chi_age', 'race4', 'birth_weight_grams', 'birthplace_city',
             'num_prev_cesarean', 'mother_date_of_birth'),
    version = 'final',
    kingco = FALSE,
    check_chi = TRUE
  ),
  output_directory = 'C:/temp/'
)

# Step 2: prep the initial results
initial_results <- etl_qa_initial_results(config = myconfig)

# Step 3: prep the final results
final_results <- etl_qa_final_results(initial_qa_results = initial_results,
                                      config = myconfig)

# Step 4: Export tables and visualizations
etl_qa_export_results(qa_results = final_results, config = myconfig)

}

}
\keyword{internal}
