% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_qa_run_pipeline.R
\name{etl_qa_final_results}
\alias{etl_qa_final_results}
\title{Final QA results for ETL QA pipeline}
\usage{
etl_qa_final_results(initial_qa_results, config)
}
\arguments{
\item{initial_qa_results}{A list containing the initial QA results.}

\item{config}{An S3 object of class "qa_data_config" containing configuration settings.}
}
\value{
A list containing formatted and combined results that consists of:
\itemize{
\item \code{missingness}: Structured summary of the proportion of missing data per variable and time point
\item \code{values}: Combined table with the frequency of categorical variables and simple statistics for numeric and date / datetime variables
\item \code{chi_standards}: Comparison of CHI (Community Health Indicator) variables values with those expected based on \code{chi_standard_varnames}
}
}
\description{
This function processes the initial results from \code{\link[=etl_qa_initial_results]{etl_qa_initial_results()}}
into a format suitable for reporting and visualization. It is the third step
run by \code{\link[=etl_qa_run_pipeline]{etl_qa_run_pipeline()}}.
}
\details{
This is an \emph{internal function} accessible only by use of \code{:::}, for example,
\code{apde.etl:::etl_qa_final_results(...)}.
}
\examples{
\dontrun{

# Step 1: generate a config object
myconfig <- etl_qa_setup_config(
  data_source_type = 'apde.data',
  data_params = list(
    function_name = 'birth',
    time_var = 'chi_year',
    time_range = c(2021, 2022),
    cols = c('chi_age', 'race4', 'birth_weight_grams', 'birthplace_city',
             'num_prev_cesarean', 'mother_date_of_birth'),
    version = 'final',
    kingco = FALSE,
    check_chi = TRUE
  ),
  output_directory = 'C:/temp/'
)

# Step 2: prep the initial results
initial_results <- etl_qa_initial_results(config = myconfig)

# Step 3: prep the final results
final_results <- etl_qa_final_results(initial_qa_results = initial_results,
                                      config = myconfig)

# Peek at the tables
head(final_results$missingness)
head(final_results$values)
head(final_results$chi_standards)

}

}
\keyword{internal}
