% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_into.R
\name{execute_copy_into}
\alias{execute_copy_into}
\title{HELPER: Execute the COPY INTO statement}
\usage{
execute_copy_into(conn, rodbc, rodbc_dsn, load_sql)
}
\arguments{
\item{conn}{Database connection}

\item{rodbc}{Whether to use RODBC}

\item{rodbc_dsn}{RODBC DSN name}

\item{load_sql}{SQL statement to execute}
}
\description{
HELPER: Execute the COPY INTO statement
}
