% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_table_check.R
\name{external_table_check}
\alias{external_table_check}
\title{Compare external table to source table to find changes}
\usage{
external_table_check(
  conn,
  db = c("inthealth_edw", "inthealth_dwhealth"),
  schema,
  table,
  db_ext = "hhs_analytics_workspace",
  conn_ext,
  schema_ext,
  table_ext,
  sql_display = TRUE,
  sql_file_path = NULL,
  overwrite = TRUE
)
}
\arguments{
\item{conn}{SQL server connection created using \code{odbc} package}

\item{db}{name of database/data warehouse for source table, must be inthealth_edw or inthealth_dwhealth}

\item{schema}{name of schema for source table}

\item{table}{name of source table}

\item{db_ext}{name of database/data warehouse for external table}

\item{conn_ext}{name of the connection to the SQL database with external table}

\item{schema_ext}{name of schema for external table}

\item{table_ext}{name of external table}

\item{sql_display}{show the SQL script in Console}

\item{sql_file_path}{write the SQL script to file}

\item{overwrite}{overwrite sql script, appends if FALSE}
}
\value{
Logical. Returns TRUE if source and external table schemas match exactly,
FALSE if they differ. Also displays/writes SQL script for external table
recreation when schemas don't match.
}
\description{
Performs schema validation between source and external tables by
querying INFORMATION_SCHEMA metadata. When inconsistencies are detected,
constructs complete external table recreation script including proper data
source references.
}
\note{
This function replaces the deprecated \code{external_table_check_f} function from the
\code{apde} package.
}
