% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_into.R
\name{get_config_param}
\alias{get_config_param}
\title{HELPER: Get parameter value from config or direct input}
\usage{
get_config_param(direct_value, table_config, server, param_name)
}
\arguments{
\item{direct_value}{Direct parameter value}

\item{table_config}{Configuration object}

\item{server}{Server name}

\item{param_name}{Parameter name in config}
}
\value{
Parameter value
}
\description{
HELPER: Get parameter value from config or direct input
}
