% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_into.R
\name{handle_table_creation}
\alias{handle_table_creation}
\title{HELPER: Handle table creation and overwrite logic}
\usage{
handle_table_creation(conn, to_schema, to_table, overwrite, table_config)
}
\arguments{
\item{conn}{Database connection}

\item{to_schema}{Target schema}

\item{to_table}{Target table}

\item{overwrite}{Whether to overwrite existing table}

\item{table_config}{Configuration object}
}
\description{
HELPER: Handle table creation and overwrite logic
}
