% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_into.R
\name{load_table_config}
\alias{load_table_config}
\title{HELPER: Load configuration from various sources}
\usage{
load_table_config(config, config_url, config_file)
}
\arguments{
\item{config}{Local config object}

\item{config_url}{Config URL}

\item{config_file}{Config file path}
}
\value{
Loaded configuration object
}
\description{
HELPER: Load configuration from various sources
}
