% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_table_from_file.R
\name{load_table_from_file}
\alias{load_table_from_file}
\title{Load a data file to a SQL table}
\usage{
load_table_from_file(
  conn,
  server = NULL,
  overall = TRUE,
  ind_yr = FALSE,
  years = NULL,
  combine_yr = FALSE,
  config = NULL,
  config_url = NULL,
  config_file = NULL,
  to_schema = NULL,
  to_table = NULL,
  server_path = NULL,
  db_name = NULL,
  azure = FALSE,
  azure_uid = NULL,
  azure_pwd = NULL,
  file_path = NULL,
  field_term = NULL,
  row_term = NULL,
  first_row = 2,
  truncate = TRUE,
  drop_index = TRUE,
  tablock = FALSE,
  test_schema = NULL,
  use_sys = FALSE
)
}
\arguments{
\item{conn}{SQL server connection created using \code{odbc} package.}

\item{server}{Name of server being used (only applies if using a YAML file).
Useful if the same table is loaded to multiple servers but with different names
or schema. Note that this is different from the \code{server_path} argument that is
used as part of the BCP command; the \code{server} argument can be any name a user
wants whereas \code{server_path} must be the actual server name.}

\item{overall}{Load a single table instead of a table for each calendar year.
Mutually exclusive with \code{ind_yr} option. Default is \code{TRUE}.}

\item{ind_yr}{Load multiple tables, one for each calendar year, with a year suffix
on each table name (e.g., mcaid_elig_2014). Mutually exclusive with \code{overall} option.
If using this option, the list of years should be provided via the \code{years} argument or
a \code{years} variable in the YAML file. Default is \code{FALSE}.}

\item{years}{Vector of individual years to make tables for (if not using YAML input).}

\item{combine_yr}{Union year-specific files into a single table. Only applies
if \code{ind_yr = TRUE}. Default is \code{FALSE}.}

\item{config}{Name of object in global environment that contains configuration
information. Use one of \code{config}, \code{config_url}, or \code{config_file}.
Should be in a YAML format with at least the following variables:
\code{to_schema}, \code{to_table}, \code{server_path}, \code{db_name}, and \code{file_path},
with possibly \code{field_term}, \code{row_term}, and \code{first_row}.
\code{to_schema} and \code{to_table}, \code{server_path}, and \code{db_name}
should all be nested under the server name if applicable, other variables
should not (but might be nested under a calendar year).}

\item{config_url}{URL of a YAML config file. Use one of \code{config}, \code{config_url}, or
\code{config_file}. Note the requirements under \code{config}.}

\item{config_file}{File path of a YAML config file. Use one of \code{config}, \code{config_url}, or
\code{config_file}. Note the requirements under \code{config}.}

\item{to_schema}{Name of the schema that data will be loaded to (if not using YAML input).}

\item{to_table}{Name of the table that data will be loaded to (if not using YAML input).}

\item{server_path}{Name of the SQL server to connect to (if not using YAML input).
If using Azure, only seems to work if you specify an existing DSN connection.}

\item{db_name}{Name of the database to use (if not using YAML input).}

\item{azure}{Flag to indicate data are being loaded to an Azure SQL server. Default is \code{FALSE}.}

\item{azure_uid}{Username for connecting to Azure Active Directory. Only use if \code{azure = TRUE}.}

\item{azure_pwd}{Password for connecting to Azure Active Directory. Only use if \code{azure = TRUE}.}

\item{file_path}{File path of data to be loaded (if not using YAML input). If
\code{ind_yr = TRUE}, this should be a named vector with the format
\code{c("2014" = "//path1/folder1/file1.ext", "2015" = "//path1/folder1/file2.ext")}
where the name matches the years to load.}

\item{field_term}{Field terminator in the data (if not using YAML input). If
using \code{ind_yr = TRUE} and the terminator differs between calendar years, this should
be a named vector with the format \code{c("overall" = "\\\\t", "2014" = "\\\\|", "2016" = "\\\\0")}
where "overall" supplies the default terminator and the other names match the years
that differ. Do not use a named vector if \code{overall = TRUE} or if there is no variation
between years. The BCP default is \verb{\\\\t}.}

\item{row_term}{Row terminator in the data (if not using YAML input). If
using \code{ind_yr = TRUE} and the terminator differs between calendar years, this should
be a named vector with the format \code{c("overall" = "\\\\n", "2014" = "\\\\r", "2016" = "\\\\0")}
where "overall" supplies the default terminator and the other names match the years
that differ. Do not use a named vector if \code{overall = TRUE} or if there is no variation
between years. The BCP default is \verb{\\\\n}.}

\item{first_row}{Row number of the first line of data (if not using YAML input).
Default is \code{2} (assumes a header row). Currently this must be the same for all years.}

\item{truncate}{Truncate existing table prior to loading. Default is \code{TRUE}.}

\item{drop_index}{Drop any existing indices prior to loading data. This can speed
loading times substantially. Use \code{add_index} to restore the index after. Default is \code{TRUE}.}

\item{tablock}{Logical (\code{TRUE} | \code{FALSE}). Lock the entire table for duration of
loading process to improve performance. Default is \code{FALSE}.}

\item{test_schema}{Write to a temporary/development schema when testing out table creation.
Will use the \code{to_schema} (specified or in the YAML file) to make a new table name of
\verb{\{to_schema\}_\{to_table\}}. Schema must already exist in the database. Most useful
when the user has an existing YAML file and does not want to overwrite it.
Only 1,000 rows will be loaded to each table. Default is \code{NULL}.}

\item{use_sys}{If the \code{sys} package is installed, use this to call BCP and see a more
informative interface in the R console. Helpful for debugging when BCP doesn't work.
Default is \code{FALSE}.}
}
\value{
This function is called for its side effects (loading data to SQL table).
It does not return a value.
}
\description{
\code{load_table_from_file()} loads a file data to a SQL table using
specified variables or a YAML config file.
}
\details{
This function loads a data file to an already existing SQL table using
specified variables or a YAML configuration file. The function is essentially a
wrapper for the SQL \href{https://learn.microsoft.com/en-us/sql/tools/bcp-utility?}{bulk copy program (BCP)}
utility using integrated authorization.

Users can specify some input functions (e.g., \code{to_table}) and rely on the config file
for the rest of the necessary information.
For all arguments that could be specified or come from a YAML file, the hierarchy is
specified > argument under server in YAML > argument not under server in YAML.
Note that arguments that should not vary between servers (e.g., \code{row_terminator})
should not be listed under a server in the YAML file.
}
\note{
This function replaces the deprecated \code{load_table_from_file()} function from the
\code{apde} package.

\strong{Note}: This function does not work to load to an Azure SQL DB (as of May 2022)
\subsection{Example YAML file with no server or individual years}{

(Assume the indentation is appropriate)

\if{html}{\out{<div class="sourceCode yaml">}}\preformatted{to_schema: raw
to_table: mcaid_elig
# optional other components like a qa_schema and qa_table, index name, vars, etc.
server_path: KCITSQLABCD51
db_name: PHClaims
file_path: //path123/importdata/Data/kc_elig_20210519.txt
field_term: \\t
row_term: \\n
}\if{html}{\out{</div>}}
}

\subsection{Example YAML file with servers (phclaims, hhsaw) and individual years}{

(Assume the indentation is appropriate)

\if{html}{\out{<div class="sourceCode yaml">}}\preformatted{phclaims:
    to_schema: raw
    to_table: mcaid_elig
    server_path: KCITSQLABCD51
    db_name: PHClaims
hhsaw:
    to_schema: raw
    to_table: mciad_elig
    server_path: kcitazxyz20.database.windows.net
    db_name: hhs_analytics_workspace
# optional other components like a qa_schema and qa_table, index name, vars, etc.
field_term: \\t
row_term: \\n
years:
    2014
    2015
    2016
2014:
    file_path: //path123/importdata/Data/kc_elig_2014.txt
    field_term: \\|
    row_term: \\r
2015:
    file_path: //path123/importdata/Data/kc_elig_2015.txt
2016:
    file_path: //path123/importdata/Data/kc_elig_2016.txt
    field_term: \\0
    row_term: \\0
}\if{html}{\out{</div>}}
}
}
\examples{
\dontrun{
load_table(conn = db_claims, server = "hhsaw", config = load_config)
myurl = "https://raw.githubusercontent.com/PHSKC-APDE/claims_data/load_mcaid_raw.yaml"
load_table(conn = db_claims, server = "phclaims",
           config_url = myurl,
           overall = FALSE,
           ind_yr = TRUE)
}

}
