% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_into.R
\name{validate_config_file}
\alias{validate_config_file}
\title{HELPER: Validate configuration file}
\usage{
validate_config_file(config_file)
}
\arguments{
\item{config_file}{Path to configuration file}
}
\description{
HELPER: Validate configuration file
}
