% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_into.R
\name{validate_config_inputs}
\alias{validate_config_inputs}
\title{HELPER: Validate configuration inputs}
\usage{
validate_config_inputs(config, config_url, config_file)
}
\arguments{
\item{config}{Local config object}

\item{config_url}{Config URL}

\item{config_file}{Config file path}
}
\description{
HELPER: Validate configuration inputs
}
