% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_into.R
\name{validate_numeric_params}
\alias{validate_numeric_params}
\title{HELPER: Validate numeric parameters}
\usage{
validate_numeric_params(max_errors, first_row)
}
\arguments{
\item{max_errors}{Maximum errors parameter}

\item{first_row}{First row parameter}
}
\description{
HELPER: Validate numeric parameters
}
