% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_into.R
\name{validate_server}
\alias{validate_server}
\title{HELPER: Validate server parameter}
\usage{
validate_server(server)
}
\arguments{
\item{server}{Server name to validate}
}
\value{
Validated server name or NA
}
\description{
HELPER: Validate server parameter
}
