% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_into.R
\name{validate_table_config}
\alias{validate_table_config}
\title{HELPER: Validate loaded configuration}
\usage{
validate_table_config(table_config)
}
\arguments{
\item{table_config}{Loaded configuration object}
}
\description{
HELPER: Validate loaded configuration
}
