% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apde.etl-package.R
\docType{package}
\name{apde.etl-package}
\alias{apde.etl}
\alias{apde.etl-package}
\title{apde.etl: APDE Tools for ETL (Extract, Transform, & Load) Procesess}
\description{
Provides tools for ETL and informatics processes.
}
\author{
\strong{Maintainer}: Danny Colombara \email{dcolombara@kingcounty.gov}

}
\keyword{internal}
