% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_qa_run_pipeline.R
\name{etl_qa_initial_results}
\alias{etl_qa_initial_results}
\title{Initial QA results for ETL QA pipeline}
\usage{
etl_qa_initial_results(config)
}
\arguments{
\item{config}{An S3 object of class "qa_data_config" containing configuration settings.}
}
\value{
A list of raw analytic output. The table structure may differ slightly depending on the original data_source. The list items include:
\itemize{
\item \code{missing_data}: The proportion of missing data for each variable and time point
\item \code{vals_continuous}: The minimum, median, mean, and maximum for all numeric variables with >= your specified \code{distinct_threshold} unique values
\item \code{vals_date}: The minimum, median, and maximum for all date / datetime variables with >= your specified \code{distinct_threshold} unique values
\item \code{vals_categorical}: A frequency table of the top 8 most frequent values
of categorical variable (and numerics or dates with < your specified \code{distinct_threshold} distinct values) PLUS a rows for \code{NA}
PLUS a row for all 'Other values'
\item \code{chi_standards}: Comparison of CHI (Community Health Indicator) variables values with those expected based on \code{rads.data::misc_chi_byvars}
}
}
\description{
This function performs the core analysis for the ETL QA pipeline, processing
data based on the provided configuration. It is the second step run by
\code{\link[=etl_qa_run_pipeline]{etl_qa_run_pipeline()}}.
}
\details{
This is an \emph{internal function} accessible only by use of \code{:::}, for example,
\code{apde.etl:::etl_qa_initial_results(...)}.
}
\examples{
\dontrun{

# Step 1: generate a config object
myconfig <- etl_qa_setup_config(
  data_source_type = 'rads',
  data_params = list(
    function_name = 'get_data_birth',
    time_var = 'chi_year',
    time_range = c(2021, 2022),
    cols = c('chi_age', 'race4', 'birth_weight_grams', 'birthplace_city',
             'num_prev_cesarean', 'mother_date_of_birth'),
    version = 'final',
    kingco = FALSE,
    check_chi = FALSE
  ),
  output_directory = 'C:/temp/'
)


# Step 2: perform the calculations
initial_results <- etl_qa_initial_results(myconfig)

# Peek at the tables
head(initial_results$missing_data)
head(initial_results$vals_categorical)
head(initial_results$vals_continuous)
head(initial_results$vals_date)

}

}
\keyword{internal}
